src/psic/bookingsystem/Room.java
/*
 * To change this license header, choose License Headers in Project Properties.
 * To change this template file, choose Tools | Templates
 * and open the template in the editor.
 */
package psic.bookingsystem;

/**
 *
 * @author Hart
 */
public class Room {
    private String name;
    
    public Room(String name) {
        this.setName(name);
    }
    public Room() {
    }

    /**
     * @return the name
     */
    public String getName() {
        return name;
    }

    /**
     * @param name the name to set
     */
    public void setName(String name) {
        this.name = name;
    }
    
}
